Restructure frontend with components and styles, create Home/Login/Signup.js, import react-router-dom
